Fixed warning

+++ stratagus/trunk/src/editor/editloop.cpp
@@ -1626,7 +1626,7 @@ local void EditorCallbackMouse(int x, int y)
 	SetStatusLine("Tile mode");
 	return;
     }
-    for (i = 0; i < sizeof(TheUI.Buttons) / sizeof(*TheUI.Buttons); ++i) {
+    for (i = 0; i < (int)(sizeof(TheUI.Buttons)/sizeof(*TheUI.Buttons)); ++i) {
 	if (x < TheUI.Buttons[i].X
 		|| x > TheUI.Buttons[i].X + TheUI.Buttons[i].Width
 		|| y < TheUI.Buttons[i].Y
